pyxdg: extend fuzzer and add corpus+dict (#7674)
* pyxdg: extend fuzzer and add corpus+dict

The extension is mainly to force xml checking before its passed into the
pyxdg library. The problem I had was any code in xml would not be
instrumented, so coverage could not be used to explore correct xml
files. I did a minor guard to use lxml instead to check xml validity, as
this could be instrumented proper. Gains are significant.

* try to fix fuzzer not part of pyxdg

* nit

* nit

* fix build failure