spdk: upgrade pip (#7915)
For some reason, there's a local python3 installation on the
oss-fuzz-base image with a pretty old pip version, 19.2.3, which doesn't
have up-to-date wheel platform tags, so, when installing SPDK
dependencies, it tries to compile them instead of using the wheels. The
compilation fails on the grpcio package for some reason.

So, to resolve this, upgrade the local pip version when building the
SPDK image, so that it's able to fetch the pre-compiled packages.  This
is a workaround for #7914.